+++ products_api/__init__.py
@@ -1,0 +1,7 @@
+from flask import Flask
+
+app = Flask(__name__)
+
+from products_api import routes
+
+app.run(debug=True)
