+++ products_api/templates/hello.html
@@ -1,0 +1,17 @@
+
+{% extends "layout.html" %}
+
+{% block title %}
+  Hello
+{% endblock %}
+
+{% block content %}
+  {% if name %}
+    <h1>Hello, {{ name }}!</h1>
+  {% else %}
+    <h1>Hello World</h1>
+  {% endif %}
+{% endblock %}
+
+{% block footer %}
+{% endblock %}
