+++ products_api/templates/index.html
@@ -1,0 +1,17 @@
+{% extends "layout.html" %}
+
+{% block title %}
+  Products API (Flask)
+{% endblock %}
+
+{% block content %}
+  <h1>Products API (Flask)</h1>
+
+  <p>Welcome to the Products API.</p>
+
+  <h2>Endpoints</h2>
+  <ul>
+    <li><a href="products">List Products</a></li>
+    <li><a href="products/1">Show Product #1</a></li>
+  </ul>
+{% endblock %}
