+++ setup.py
@@ -1,0 +1,10 @@
+from setuptools import setup
+
+setup(
+    name='products_api',
+    packages=['products_api'],
+    include_package_data=True,
+    install_requires=[
+        'flask',
+    ],
+)
